Major updates: - Initialized Firebase with DefaultFirebaseOptions - Created LoginScreen with email/password validation and Firebase auth - Added religion-based redirection after login (Christianity / Islam) - Implemented MainNavigation with bottom tabs and user name display - Built ProfileScreen with user info, profile photo, and logout - Developed EditProfileScreen for profile updates including photo upload - Improved UX with loading indicators and SnackBar feedback - Updated navigation to use named routes for login and register
