+++ pubspec.yaml
@@ -11,23 +11,29 @@ dependencies:
   flutter:
     sdk: flutter
 
-  firebase_core: ^3.14.0
-  firebase_auth: ^5.6.0
-  cloud_firestore: ^5.6.9
+  image_picker: ^1.1.2
+  firebase_auth: ^5.6.2
+  firebase_analytics: ^11.0.1
+  firebase_crashlytics: ^4.3.9
 
-  timezone: ^0.9.2
+  firebase_core: ^3.15.1
+
+  cloud_firestore: ^5.6.11
+  firebase_storage: ^12.4.9
+
+  timezone: ^0.10.1
 
   cupertino_icons: ^1.0.8
   
   provider: ^6.1.2
   flutter_spinkit: ^5.2.0
-  google_fonts: ^6.2.0 # ✅ You can still keep this for future
-  flutter_local_notifications: ^18.0.1
+  google_fonts: ^6.2.0
+  flutter_local_notifications: ^19.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.0
+  flutter_lints: ^6.0.0
 
 flutter:
   uses-material-design: true
@@ -35,10 +41,9 @@ flutter:
   assets:
     - assets/images/
 
-  
-  # fonts:
-  #   - family: Poppins
-  #     fonts:
-  #       - asset: assets/fonts/Poppins-Regular.ttf
-  #       - asset: assets/fonts/Poppins-Bold.ttf
-  #         weight: 700
+# fonts:
+#   - family: Poppins
+#     fonts:
+#       - asset: assets/fonts/Poppins-Regular.ttf
+#       - asset: assets/fonts/Poppins-Bold.ttf
+#         weight: 700
